Rename `vue/no-setup-props-destructure` to `vue/no-setup-props-reactivity-loss` and remove from config
