+++ docs/rules/no-setup-props-reactivity-loss.md
@@ -1,21 +1,20 @@
 ---
 pageClass: rule-details
 sidebarDepth: 0
-title: vue/no-setup-props-destructure
+title: vue/no-setup-props-reactivity-loss
 description: disallow usages that lose the reactivity of `props` passed to `setup`
-since: v7.0.0
 ---
-# vue/no-setup-props-destructure
+# vue/no-setup-props-reactivity-loss
 
 > disallow usages that lose the reactivity of `props` passed to `setup`
 
-- :no_entry_sign: This rule was **deprecated** and replaced by [vue/no-setup-props-reactivity-loss](no-setup-props-reactivity-loss.md) rule.
+- :exclamation: <badge text="This rule has not been released yet." vertical="middle" type="error"> ***This rule has not been released yet.*** </badge>
 
 ## :book: Rule Details
 
-This rule reports the destructuring of `props` passed to `setup` causing the value to lose reactivity.
+This rule reports the destructuring or member expression of `props` passed to `setup` causing the value to lose reactivity.
 
-<eslint-code-block :rules="{'vue/no-setup-props-destructure': ['error']}">
+<eslint-code-block :rules="{'vue/no-setup-props-reactivity-loss': ['error']}">
 
 ```vue
 <script>
@@ -38,7 +37,7 @@ export default {
 
 Destructuring the `props` passed to `setup` will cause the value to lose reactivity.
 
-<eslint-code-block :rules="{'vue/no-setup-props-destructure': ['error']}">
+<eslint-code-block :rules="{'vue/no-setup-props-reactivity-loss': ['error']}">
 
 ```vue
 <script>
@@ -61,7 +60,7 @@ export default {
 
 Also, destructuring in root scope of `setup()` should error, but ok inside nested callbacks or returned render functions:
 
-<eslint-code-block :rules="{'vue/no-setup-props-destructure': ['error']}">
+<eslint-code-block :rules="{'vue/no-setup-props-reactivity-loss': ['error']}">
 
 ```vue
 <script>
@@ -97,11 +96,7 @@ Nothing.
 - [Guide - Composition API - Setup](https://vuejs.org/api/composition-api-setup.html)
 - [Vue RFCs - 0013-composition-api](https://github.com/vuejs/rfcs/blob/master/active-rfcs/0013-composition-api.md)
 
-## :rocket: Version
-
-This rule was introduced in eslint-plugin-vue v7.0.0
-
 ## :mag: Implementation
 
-- [Rule source](https://github.com/vuejs/eslint-plugin-vue/blob/master/lib/rules/no-setup-props-destructure.js)
-- [Test source](https://github.com/vuejs/eslint-plugin-vue/blob/master/tests/lib/rules/no-setup-props-destructure.js)
+- [Rule source](https://github.com/vuejs/eslint-plugin-vue/blob/master/lib/rules/no-setup-props-reactivity-loss.js)
+- [Test source](https://github.com/vuejs/eslint-plugin-vue/blob/master/tests/lib/rules/no-setup-props-reactivity-loss.js)
