+++ lib/index.js
@@ -131,6 +131,7 @@ module.exports = {
     'no-restricted-v-bind': require('./rules/no-restricted-v-bind'),
     'no-root-v-if': require('./rules/no-root-v-if'),
     'no-setup-props-destructure': require('./rules/no-setup-props-destructure'),
+    'no-setup-props-reactivity-loss': require('./rules/no-setup-props-reactivity-loss'),
     'no-shared-component-data': require('./rules/no-shared-component-data'),
     'no-side-effects-in-computed-properties': require('./rules/no-side-effects-in-computed-properties'),
     'no-spaces-around-equal-signs-in-attribute': require('./rules/no-spaces-around-equal-signs-in-attribute'),
