+++ lib/rules/no-setup-props-destructure.js
@@ -3,269 +3,22 @@
  * See LICENSE file in root directory for full license.
  */
 'use strict'
-const { findVariable } = require('@eslint-community/eslint-utils')
-const utils = require('../utils')
+const baseRule = require('./no-setup-props-reactivity-loss')
 
 module.exports = {
+  // eslint-disable-next-line eslint-plugin/require-meta-schema, eslint-plugin/prefer-message-ids, internal/no-invalid-meta, eslint-plugin/require-meta-type -- inherit schema from base rule
   meta: {
-    type: 'suggestion',
+    ...baseRule.meta,
+    // eslint-disable-next-line eslint-plugin/require-meta-docs-description, internal/no-invalid-meta-docs-categories, eslint-plugin/meta-property-ordering
     docs: {
-      description: 'disallow destructuring of `props` passed to `setup`',
-      categories: ['vue3-essential', 'essential'],
+      ...baseRule.meta.docs,
       url: 'https://eslint.vuejs.org/rules/no-setup-props-destructure.html'
     },
-    fixable: null,
-    schema: [],
-    messages: {
-      destructuring:
-        'Destructuring the `props` will cause the value to lose reactivity.',
-      getProperty:
-        'Getting a value from the `props` in root scope of `{{scopeName}}` will cause the value to lose reactivity.'
-    }
+    deprecated: true,
+    replacedBy: ['no-setup-props-reactivity-loss']
   },
-  /**
-   * @param {RuleContext} context
-   * @returns {RuleListener}
-   **/
+  /** @param {RuleContext} context */
   create(context) {
-    /**
-     * @typedef {object} ScopePropsReferences
-     * @property {Set<Identifier>} refs
-     * @property {string} scopeName
-     */
-    /** @type {Map<FunctionDeclaration | FunctionExpression | ArrowFunctionExpression | Program, ScopePropsReferences>} */
-    const setupScopePropsReferenceIds = new Map()
-    const wrapperExpressionTypes = new Set([
-      'ArrayExpression',
-      'ObjectExpression'
-    ])
-
-    /**
-     * @param {ESNode} node
-     * @param {string} messageId
-     * @param {string} scopeName
-     */
-    function report(node, messageId, scopeName) {
-      context.report({
-        node,
-        messageId,
-        data: {
-          scopeName
-        }
-      })
-    }
-
-    /**
-     * @param {Pattern} left
-     * @param {Expression | null} right
-     * @param {ScopePropsReferences} propsReferences
-     */
-    function verify(left, right, propsReferences) {
-      if (!right) {
-        return
-      }
-
-      const rightNode = utils.skipChainExpression(right)
-
-      if (
-        wrapperExpressionTypes.has(rightNode.type) &&
-        isPropsMemberAccessed(rightNode, propsReferences)
-      ) {
-        return report(rightNode, 'getProperty', propsReferences.scopeName)
-      }
-
-      if (
-        left.type !== 'ArrayPattern' &&
-        left.type !== 'ObjectPattern' &&
-        rightNode.type !== 'MemberExpression'
-      ) {
-        return
-      }
-
-      /** @type {Expression | Super} */
-      let rightId = rightNode
-      while (rightId.type === 'MemberExpression') {
-        rightId = utils.skipChainExpression(rightId.object)
-      }
-      if (rightId.type === 'Identifier' && propsReferences.refs.has(rightId)) {
-        report(left, 'getProperty', propsReferences.scopeName)
-      }
-    }
-
-    /**
-     * @param {Expression} node
-     * @param {ScopePropsReferences} propsReferences
-     */
-    function isPropsMemberAccessed(node, propsReferences) {
-      const propRefs = [...propsReferences.refs.values()]
-
-      return propRefs.some((props) => {
-        const isPropsInExpressionRange = utils.inRange(node.range, props)
-        const isPropsMemberExpression =
-          props.parent.type === 'MemberExpression' &&
-          props.parent.object === props
-
-        return isPropsInExpressionRange && isPropsMemberExpression
-      })
-    }
-
-    /**
-     * @typedef {object} ScopeStack
-     * @property {ScopeStack | null} upper
-     * @property {FunctionDeclaration | FunctionExpression | ArrowFunctionExpression | Program} scopeNode
-     */
-    /**
-     * @type {ScopeStack | null}
-     */
-    let scopeStack = null
-
-    /**
-     * @param {Pattern | null} node
-     * @param {FunctionDeclaration | FunctionExpression | ArrowFunctionExpression | Program} scopeNode
-     * @param {string} scopeName
-     */
-    function processPattern(node, scopeNode, scopeName) {
-      if (!node) {
-        // no arguments
-        return
-      }
-      if (
-        node.type === 'RestElement' ||
-        node.type === 'AssignmentPattern' ||
-        node.type === 'MemberExpression'
-      ) {
-        // cannot check
-        return
-      }
-      if (node.type === 'ArrayPattern' || node.type === 'ObjectPattern') {
-        report(node, 'destructuring', scopeName)
-        return
-      }
-
-      const variable = findVariable(context.getScope(), node)
-      if (!variable) {
-        return
-      }
-      const propsReferenceIds = new Set()
-      for (const reference of variable.references) {
-        // If reference is in another scope, we can't check it.
-        if (reference.from !== context.getScope()) {
-          continue
-        }
-
-        if (!reference.isRead()) {
-          continue
-        }
-
-        propsReferenceIds.add(reference.identifier)
-      }
-      setupScopePropsReferenceIds.set(scopeNode, {
-        refs: propsReferenceIds,
-        scopeName
-      })
-    }
-    return utils.compositingVisitors(
-      {
-        /**
-         * @param {FunctionExpression | FunctionDeclaration | ArrowFunctionExpression | Program} node
-         */
-        'Program, :function'(node) {
-          scopeStack = {
-            upper: scopeStack,
-            scopeNode: node
-          }
-        },
-        /**
-         * @param {FunctionExpression | FunctionDeclaration | ArrowFunctionExpression | Program} node
-         */
-        'Program, :function:exit'(node) {
-          scopeStack = scopeStack && scopeStack.upper
-
-          setupScopePropsReferenceIds.delete(node)
-        },
-        /**
-         * @param {CallExpression} node
-         */
-        CallExpression(node) {
-          if (!scopeStack) {
-            return
-          }
-
-          const propsReferenceIds = setupScopePropsReferenceIds.get(
-            scopeStack.scopeNode
-          )
-
-          if (!propsReferenceIds) {
-            return
-          }
-
-          if (isPropsMemberAccessed(node, propsReferenceIds)) {
-            report(node, 'getProperty', propsReferenceIds.scopeName)
-          }
-        },
-        /**
-         * @param {VariableDeclarator} node
-         */
-        VariableDeclarator(node) {
-          if (!scopeStack) {
-            return
-          }
-          const propsReferenceIds = setupScopePropsReferenceIds.get(
-            scopeStack.scopeNode
-          )
-          if (!propsReferenceIds) {
-            return
-          }
-          verify(node.id, node.init, propsReferenceIds)
-        },
-        /**
-         * @param {AssignmentExpression} node
-         */
-        AssignmentExpression(node) {
-          if (!scopeStack) {
-            return
-          }
-          const propsReferenceIds = setupScopePropsReferenceIds.get(
-            scopeStack.scopeNode
-          )
-          if (!propsReferenceIds) {
-            return
-          }
-          verify(node.left, node.right, propsReferenceIds)
-        }
-      },
-      utils.defineScriptSetupVisitor(context, {
-        onDefinePropsEnter(node) {
-          let target = node
-          if (
-            target.parent &&
-            target.parent.type === 'CallExpression' &&
-            target.parent.arguments[0] === target &&
-            target.parent.callee.type === 'Identifier' &&
-            target.parent.callee.name === 'withDefaults'
-          ) {
-            target = target.parent
-          }
-          if (!target.parent) {
-            return
-          }
-
-          /** @type {Pattern|null} */
-          let id = null
-          if (target.parent.type === 'VariableDeclarator') {
-            id = target.parent.init === target ? target.parent.id : null
-          } else if (target.parent.type === 'AssignmentExpression') {
-            id = target.parent.right === target ? target.parent.left : null
-          }
-          processPattern(id, context.getSourceCode().ast, '<script setup>')
-        }
-      }),
-      utils.defineVueVisitor(context, {
-        onSetupFunctionEnter(node) {
-          const propsParam = utils.skipDefaultParamValue(node.params[0])
-          processPattern(propsParam, node, 'setup()')
-        }
-      })
-    )
+    return baseRule.create(context)
   }
 }
